==> 3_Acro_Pocillo_PCA.Rmd <==
---
title: "3_Acro_Pocillo_PCA"
author: "Magena Marzonie"
date: "09/02/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Load libraries 
```{r}
library(vegan)
library(ggsci)
library(tidyverse)
library(kmer)
library(phangorn)
library(GUniFrac)
library(ggtree)
library(patchwork)
library(bioseq)
library(ape)
library(Biostrings)
select = dplyr::select
```

#Load data 
```{r}
#load environmental data 
load("Data/SiteDisturbanceHistory_DHW.RData")
load("Data/SiteDisturbanceHistory_DHW.RData")

#load seq data from 2_UPGMA
load("Data/du_acro.RData")
load("Data/du_poci.RData")
load("Data/du_pver.Rdata")
load("Data/du_pmea.Rdata")
load("Data/du_punk.Rdata")
load("Data/full_meta.RData")
load("Data/full_meta2.RData")
```

#Data Wrangle 
```{r}
#Combining symbiont unifrac distances and environmental metadata
full_meta2 = full_meta2 %>% 
  mutate(DHW = as.numeric(DHW),
         Reef = as.factor(Reef.x),
         Site = as.factor(Site),
         Exposure = as.factor(Exposure),
         catBleaching = as.factor(catBleaching),
         Aspect = as.factor(Aspect))
```

# ----------------------------------------------------------
Permanova/ PCoA: **Pocillopora verrucosa**
# ----------------------------------------------------------

Workflow for each host species: 
1) Run full model (dbRDA) and plot correlation matrix of numeric variables 
2) Consult VIF scores of full model and correlation matrix. Narrow down variables to remove any obvious collinear variables 
3) Select environmental variables of interest with reference to correlation plots 
4) Run backward stepwise model (ordistep) to reduce model
5) Plot constrained ordistep (dbRDA) using PCoA


**P. verrucosa**
```{r}
#importing unifrac distances for analysis 
dist_pver <- as.dist(du_pver, diag = FALSE)
dim(du_pver)

adonis_meta_pver <- full_meta2 %>%
  dplyr::select(plate_position, Sector, Reef.x, Site, Species, 
                catBleaching, Depth, DHW2020 = `DHW`, GPS_S = `GPS south`, 
                GPS_E = `GPS east`, Aspect, Exposure, maxDHW, 
                meanDHW, recent.maxDHW, recent.meanDHW, DHW2, 
                DHW3, DHW4, DHW6, DHW8, DHW9, returnDHW3, returnDHW4, 
                returnDHW6, rangeSST, varSST, MMM) %>%
  filter(plate_position %in% rownames(du_pver)) %>%
  tibble::column_to_rownames(var = "plate_position") %>%
  mutate(sample_name = rownames(.))

# NOTE: Need to make sure there are no alias factors. E.g. exposure is an alias of Reef (e.g. all Marion = Exposed, Site and Reef contain aliases, DHW2 == DHW3, etc)


# Make a correlation matrix of numeric variables
cm_pver <- cor(adonis_meta_pver %>% select(Depth, maxDHW, meanDHW, recent.maxDHW, 
                                      recent.meanDHW, DHW3,  DHW6, DHW8, 
                                      DHW9, rangeSST, varSST, MMM, GPS_S, DHW2020))
library(corrplot)
corrplot(cm_pver) # Clearly a lot of the DHW terms are positively related and approaching redundant

# Check the vif scores of the full model
ord_pver_full <- dbrda(dist_pver ~  Depth + maxDHW + catBleaching + DHW2020 + meanDHW + recent.maxDHW + recent.meanDHW + DHW3 + DHW4 + DHW6 + DHW8 + DHW9 + rangeSST + varSST + MMM + GPS_S + GPS_E + DHW2020 + catBleaching, data=adonis_meta_pver)
sort(vif.cca(ord_pver_full)) 
# according to vif.cca documentation, values over 10 indicate redundant constraints
    #      Depth           DHW6           DHW3  recent.maxDHW           DHW8         maxDHW           DHW9           DHW4        meanDHW       rangeSST 
    #   1.513119       8.139152       9.269449      10.521798      10.882881      15.392102      21.143380      30.863009      58.925047      65.211384 
    #     varSST          GPS_S recent.meanDHW            MMM 
    # 134.792127     188.566323     190.258298     418.940058

# Reduce the model (consult the corplot and vif scores)
ord_pver <- dbrda(dist_pver ~ GPS_S + Depth + DHW3 + DHW6 + recent.maxDHW + catBleaching + varSST, data=adonis_meta_pver)

# Re check the new vif scores
sort(vif.cca(ord_pver))
# catBleaching         Depth          DHW3 recent.maxDHW        varSST          DHW6         GPS_S 
#      1.034575      1.123910      1.428661      1.809345      3.699175      8.465436     10.678142 
#       DHW2020 
#     14.428485 

# Use ordistep to further refine the model
os_pver_backward <- ordistep(ord_pver, direction = "backward", trace = FALSE, permutations = 9999)
anova(os_pver_backward, by = 'margin')
plot(os_pver_backward)

#extract scores
pver_scores <- as.data.frame(scores(os_pver_backward, display = "sites")) %>%
  tibble::rownames_to_column(var = "plate_position") %>%
  left_join(., adonis_meta_pver %>% tibble::rownames_to_column(var = "plate_position"))

#extract vectors
pver_vectors <- as.data.frame(os_pver_backward$CCA$biplot) %>%
  tibble::rownames_to_column(var = "factors")

#plot PCoA
g1 = ggplot(pver_scores, aes(x = dbRDA1, y = dbRDA2)) +
  geom_point(aes(fill = Reef.x), size = 4, shape = 21) +
    facet_wrap(~Reef.x) +
  #geom_encircle(aes(group = k), fill = "grey50", s_shape = 1, expand = 0, alpha = 0.2, show.legend = FALSE) +
  #geom_text(data = pver_vectors, aes(x = dbRDA1, y = dbRDA2, label = factors), size = 3) +
  geom_segment(data = pver_vectors, aes(x = 0, xend = dbRDA1, y = 0, yend = dbRDA2), size = 0.8, arrow = arrow(length = unit(0.25, "cm")), colour = "grey") +
  theme(legend.position = "none", aspect.ratio = 1) 
  #scale_fill_viridis_d(option = "magma")

g1
```




#Variance partitioning to extract individual contributions. This function uses adjusted R-squared in dbRDA. Collinear variables are allowed to be tested. 
```{r}
varp = varpart(dist_pver, ~ Depth, ~DHW3, ~GPS_S, data = adonis_meta_pver)

plot(varp, digits = 2, Xnames = c("Depth", "DHW3", "GPS_S"), bg = c("navy", "tomato", "orange"))
```
**For P. verrucosa symbionts, environmental drivers tested are not explaining much variation (residuals = 0.96). What else could be driving P. verrucosa communities? It does not seem that we are picking up on it ** 

**P. meandrina** 
```{r}
#importing unifrac distances for analysis 
dist_pmea <- as.dist(du_pmea, diag = FALSE)
dim(du_pmea)

adonis_meta_pmea <- full_meta2 %>%
  dplyr::select(plate_position, Sector, Reef.x, Site, Species, 
                catBleaching, Depth, DHW2020 = `DHW`, catBleaching, GPS_S = `GPS south`, 
                GPS_E = `GPS east`, Aspect, Exposure, maxDHW, 
                meanDHW, recent.maxDHW, recent.meanDHW, DHW2, 
                DHW3, DHW4, DHW6, DHW8, DHW9, returnDHW3, returnDHW4, 
                returnDHW6, rangeSST, varSST, MMM) %>%
  filter(plate_position %in% rownames(du_pmea)) %>%
  tibble::column_to_rownames(var = "plate_position") %>%
  mutate(sample_name = rownames(.))

# NOTE: Need to make sure there are no alias factors. E.g. exposure is an alias of Reef (e.g. all Marion = Exposed, Site and Reef contain aliases, DHW2 == DHW3, etc)

# Make a correlation matrix of numeric variables
cm <- cor(adonis_meta_pmea %>% select(Depth, maxDHW, meanDHW, recent.maxDHW, 
                                      recent.meanDHW, DHW3, DHW4, DHW6, DHW8, 
                                      DHW9, rangeSST, varSST, MMM, GPS_S, DHW2020))
library(corrplot)
corrplot(cm) # Clearly a lot of the DHW terms are positively related and approaching redundant

# Check the vif scores of the full model
ord_pmea_full <- dbrda(dist_pmea ~ Depth + maxDHW + meanDHW + recent.maxDHW + recent.meanDHW + DHW3 + DHW4 + DHW6 + DHW8 + DHW9 + rangeSST + varSST + MMM + GPS_S + DHW2020 + catBleaching, data=adonis_meta_pmea)
sort(vif.cca(ord_pmea_full)) 
# Depth           DHW4           DHW6           DHW9           DHW8         maxDHW           DHW3  recent.maxDHW        meanDHW       rangeSST recent.meanDHW         varSST 
#       1.961623       5.524548       6.432968      10.005341      10.535683      11.883727      12.147581      13.755357      56.646502      67.138168     133.177017     155.900084 
#          GPS_S            MMM 
#     216.759822     414.329321  

# Reduce the model (consult the corplot and vif scores)
ord_pmea <- dbrda(dist_pmea ~ GPS_S + Depth + DHW3 + DHW6 + recent.maxDHW + catBleaching + DHW2020 + varSST, data = adonis_meta_pmea)

# Re check the new vif scores
sort(vif.cca(ord_pmea))
 # catBleaching         Depth          DHW3 recent.maxDHW        varSST          DHW6       DHW2020         GPS_S 
 #     1.163363      1.422231      1.891411      2.247049      4.144833      5.076538     10.220901     12.225265 

# Use ordistep to further refine the model
os_pmea_backward <- ordistep(ord_pmea, direction = "backward", permutations = 999)
anova(os_pmea_backward, by = 'margin')
plot(os_pmea_backward)

pmea_scores <- as.data.frame(scores(os_pmea_backward, display = "sites")) %>%
  tibble::rownames_to_column(var = "plate_position") %>%
  left_join(., adonis_meta_pmea %>% tibble::rownames_to_column(var = "plate_position"))

pmea_vectors <- as.data.frame(os_pmea_backward$CCA$biplot) %>%
  tibble::rownames_to_column(var = "factors")

g2 = ggplot(pmea_scores, aes(x = dbRDA1, y = dbRDA2)) +
  geom_point(aes(fill = Reef.x), size = 4, shape = 21) +
  #ggalt::geom_encircle(aes(group = Reef.x), fill = "grey50", s_shape = 1, expand = 0, alpha = 0.2, show.legend = FALSE) +
  facet_wrap(~Reef.x) +
  geom_text(data = pmea_vectors, aes(x = dbRDA1, y = dbRDA2, label = factors), size = 4) +
  geom_segment(data = pmea_vectors, aes(x = 0, xend = dbRDA1, y = 0, yend = dbRDA2), size = 0.5, arrow = arrow(length = unit(0.25, "cm")), colour = "grey") +
  theme(legend.position = "right", aspect.ratio = 1) 
 # scale_fill_viridis_d(option = "magma")

g2

```

```{r}
varp_pmea = varpart(dist_pmea, ~DHW6, ~catBleaching, ~varSST, ~GPS_S, data = adonis_meta_pmea)

plot(varp_pmea, digits = 2, Xnames = c("DHW6", "catBleaching", "varSST", "GPS_S"), bg = c("navy", "tomato", "orange", "green"))
```
**There is a strong latitudinal break in the P. meandrina samples that reflects the two 'sub-groups' of symbionts present. This is highlighted by the majority of influence explained by latitude (GPS_S) in the model. GPS_S explains at least 4x as much variation as any of the other factors present, including bleaching condition of corals or variation in SST ** 



```{r}
#importing unifrac distances for analysis 
dist_punk <- as.dist(du_punk, diag = FALSE)
dim(du_punk)

adonis_meta_punk <- full_meta2 %>%
  dplyr::select(plate_position, Sector, Reef.x, Site, Species, 
                catBleaching, Depth, DHW2020 = `DHW`, GPS_S = `GPS south`, 
                GPS_E = `GPS east`, Aspect, Exposure, maxDHW, 
                meanDHW, recent.maxDHW, recent.meanDHW, DHW2, 
                DHW3, DHW4, DHW6, DHW8, DHW9, returnDHW3, returnDHW4, 
                returnDHW6, rangeSST, varSST, MMM) %>%
  filter(plate_position %in% rownames(du_punk)) %>%
  tibble::column_to_rownames(var = "plate_position") %>%
  mutate(sample_name = rownames(.))

# NOTE: Need to make sure there are no alias factors. E.g. exposure is an alias of Reef (e.g. all Marion = Exposed, Site and Reef contain aliases, DHW2 == DHW3, etc)


# Make a correlation matrix of numeric variables
cm_punk <- cor(adonis_meta_punk %>% select(Depth, maxDHW, meanDHW, recent.maxDHW, 
                                      recent.meanDHW, DHW3, DHW4, DHW6, DHW8, 
                                      DHW9, rangeSST, varSST, MMM, GPS_S, DHW2020))
library(corrplot)
corrplot(cm_punk) # Clearly a lot of the DHW terms are positively related and approaching redundant

# Check the vif scores of the full model
ord_punk_full <- dbrda(dist_punk ~  Depth + maxDHW + meanDHW + recent.maxDHW + recent.meanDHW + DHW3 + DHW4 + DHW6 + DHW8 + DHW9 + rangeSST + varSST + MMM + GPS_S + catBleaching + DHW2020, data=adonis_meta_punk)
sort(vif.cca(ord_punk_full)) 
# according to vif.cca documentation, values over 10 indicate redundant constraints
    #  Depth           DHW6           DHW8           DHW3           DHW4           DHW9 
    #   1.602809       5.019220       6.170613       7.151695       8.330065      13.456450 
    #     maxDHW  recent.maxDHW        meanDHW recent.meanDHW       rangeSST         varSST 
    #  15.191953      15.506837      28.891993     113.987710     155.276867     179.948250 
    #      GPS_S            MMM 
    # 735.291524     780.690427 

# Reduce the model (consult the corplot and vif scores)
ord_punk <- dbrda(dist_punk ~ GPS_S + Depth + DHW3 + DHW6 + recent.maxDHW + catBleaching + DHW2020 + varSST, data=adonis_meta_punk)

# Re check the new vif scores
sort(vif.cca(ord_punk))
     #    Depth          DHW3        varSST recent.maxDHW          DHW6         GPS_S 
     # 1.494017      2.831416      3.215707      3.781945      6.553586      6.997421 

# Use ordistep to further refine the model
os_punk_backward <- ordistep(ord_punk, direction = "backward", trace = FALSE)
anova(os_punk_backward, by = 'margin')
plot(os_punk_backward)


punk_scores <- as.data.frame(scores(os_punk_backward, display = "sites")) %>%
  tibble::rownames_to_column(var = "plate_position") %>%
  left_join(., adonis_meta_punk %>% tibble::rownames_to_column(var = "plate_position"))

punk_vectors <- as.data.frame(os_punk_backward$CCA$biplot) %>%
  tibble::rownames_to_column(var = "factors")

g3 = ggplot(punk_scores, aes(x = dbRDA1, y = dbRDA2)) +
  geom_point(aes(fill = Reef.x), size = 4, shape = 21) +
  #ggalt::geom_encircle(aes(group = Reef.x), fill = "grey50", s_shape = 1, expand = 0, alpha = 0.2, show.legend = FALSE) +
  facet_wrap(~Reef.x) +
 # geom_text(data = punk_vectors, aes(x = dbRDA1, y = dbRDA2, label = factors), size = 4) +
  geom_segment(data = punk_vectors, aes(x = 0, xend = dbRDA1, y = 0, yend = dbRDA2), size = 0.5, arrow = arrow(length = unit(0.25, "cm")), colour = "grey") +
  theme(legend.position = "none", aspect.ratio = 1) +
  scale_fill_viridis_d(option = "magma")

g3

```



```{r}
varp_punk = varpart(dist_punk, ~GPS_S, ~catBleaching, ~varSST,  data = adonis_meta_punk)

plot(varp_punk, digits = 2, Xnames = c("GPS_S", "catBleaching", "varSST"), bg = c("navy", "tomato", "orange"))
```
**Similar to P. meandrina, the unknown host symbionts are highly driven by latitude compared to other drivers. The influence of GPS_S is 3-times higher than the influence of bleaching condition ** 



**A. humilis**
```{r}
#importing unifrac distances for analysis 

#removing outlier samples from the dbRDA output. Optional step! Choose an option
dbrda_outliers <- c("Plate7_D006", "Plate6_E009", "Plate6_G009", "Plate6_D012", "Plate7_H009", "Plate7_A010", "Plate6_B012")
#dist_acro <- as.dist(du_acro, diag = FALSE)
du_acro.original = du_acro
du_acro <- du_acro[!row.names(du_acro) %in% dbrda_outliers, !colnames(du_acro) %in% dbrda_outliers]
dist_acro <- as.dist(du_acro, diag = FALSE)



adonis_meta_acro <- full_meta2 %>%
  dplyr::select(plate_position, Sector, Reef.x, Site, Species, 
                catBleaching, Depth, DHW2020 = `DHW`, GPS_S = `GPS south`, 
                GPS_E = `GPS east`, Aspect, Exposure, maxDHW, 
                meanDHW, recent.maxDHW, recent.meanDHW, DHW2, 
                DHW3, DHW4, DHW6, DHW8, DHW9, returnDHW3, returnDHW4, 
                returnDHW6, rangeSST, varSST, MMM) %>%
  filter(plate_position %in% rownames(du_acro)) %>%
  tibble::column_to_rownames(var = "plate_position") %>%
  mutate(sample_name = rownames(.))

# NOTE: Need to make sure there are no alias factors. E.g. exposure is an alias of Reef (e.g. all Marion = Exposed, Site and Reef contain aliases, DHW2 == DHW3, etc)


# Make a correlation matrix of numeric variables
cm_acro <- cor(adonis_meta_acro %>% select(Depth, maxDHW, meanDHW, recent.maxDHW, 
                                      recent.meanDHW, DHW3, DHW4, DHW6, DHW8, 
                                      DHW9, rangeSST, varSST, MMM, GPS_S, DHW2020))
library(corrplot)
corrplot(cm_acro) # Clearly a lot of the DHW terms are positively related and approaching redundant

# Check the vif scores of the full model
ord_acro_full <- dbrda(dist_acro ~  Depth + maxDHW + meanDHW + recent.maxDHW + recent.meanDHW + DHW3 + DHW4 + DHW6 + DHW8 + DHW9 + rangeSST + varSST + MMM + GPS_S + catBleaching + DHW2020, data=adonis_meta_acro)
sort(vif.cca(ord_acro_full)) 
# according to vif.cca documentation, values over 10 indicate redundant constraints
    #  Depth           DHW6           DHW8           DHW3           DHW4           DHW9 
    #   1.602809       5.019220       6.170613       7.151695       8.330065      13.456450 
    #     maxDHW  recent.maxDHW        meanDHW recent.meanDHW       rangeSST         varSST 
    #  15.191953      15.506837      28.891993     113.987710     155.276867     179.948250 
    #      GPS_S            MMM 
    # 735.291524     780.690427 

# Reduce the model (consult the corplot and vif scores)
ord_acro <- dbrda(dist_acro ~ GPS_S + Depth + DHW3 + DHW6 + recent.maxDHW + catBleaching  + varSST, data=adonis_meta_acro)

# Re check the new vif scores
sort(vif.cca(ord_acro))
  # Depth          DHW3          DHW6 recent.maxDHW        varSST         GPS_S 
  #    1.296128      1.326525      1.449396      1.653195      2.003441      2.577150 


# Use ordistep to further refine the model
os_acro_backward <- ordistep(ord_acro, direction = "backward", trace = FALSE)
anova(os_acro_backward, by = 'margin')
plot(os_acro_backward)

acro_scores <- as.data.frame(scores(os_acro_backward, display = "sites")) %>%
  tibble::rownames_to_column(var = "plate_position") %>%
  left_join(., adonis_meta_acro %>% tibble::rownames_to_column(var = "plate_position"))

acro_vectors <- as.data.frame(os_acro_backward$CCA$biplot) %>%
  tibble::rownames_to_column(var = "factors")

g4 = ggplot(acro_scores, aes(x = dbRDA1, y = dbRDA2)) +
  geom_point(aes(fill = Reef.x), size = 4, shape = 21) +
  #ggalt::geom_encircle(aes(group = Reef.x), fill = "grey50", s_shape = 1, expand = 0, alpha = 0.2, show.legend = FALSE) +
  facet_wrap(~Reef.x) +
 # geom_text(data = acro_vectors, aes(x = dbRDA1, y = dbRDA2, label = factors), size = 4) +
  geom_segment(data = acro_vectors, aes(x = 0, xend = dbRDA1, y = 0, yend = dbRDA2), size = 0.5, arrow = arrow(length = unit(0.25, "cm")), colour = "grey") +
  theme(aspect.ratio = 1) +
  theme_light()
  #scale_fill_viridis_d(option = "magma")

g4
```
```{r}
ggplot(acro_scores, aes(x = dbRDA1, y = dbRDA2)) +
  geom_text(data = acro_vectors, aes(x = dbRDA1, y = dbRDA2, label = factors), size = 4) +
  geom_segment(data = acro_vectors, aes(x = 0, xend = dbRDA1, y = 0, yend = dbRDA2), size = 0.5, arrow = arrow(length = unit(0.25, "cm")), colour = "grey") +
  theme(aspect.ratio = 1) +
  theme_light()
```

```{r}
#which samples are outliers in the acropora ordination? 
acro_outliers = acro_scores %>% filter(dbRDA2 < -1)
acro_outliers %>% pull(plate_position)

#outlier samples are from Moore, Holmes and Osprey 

acro_new = acro_scores %>% filter(dbRDA2 > -1)



g4 = ggplot(acro_scores, aes(x = dbRDA1, y = dbRDA2)) +
  geom_point(aes(fill = Reef.x), size = 4, shape = 21) +
  #ggalt::geom_encircle(aes(group = Reef.x), fill = "grey50", s_shape = 1, expand = 0, alpha = 0.2, show.legend = FALSE) +
  #facet_wrap(~Reef.x) +
  geom_text(data = acro_vectors, aes(x = dbRDA1, y = dbRDA2, label = factors), size = 4) +
  geom_segment(data = acro_vectors, aes(x = 0, xend = dbRDA1, y = 0, yend = dbRDA2), size = 0.5, arrow = arrow(length = unit(0.25, "cm")), colour = "grey") +
  theme(legend.position = "none", aspect.ratio = 1) + 
  scale_fill_viridis_d(option = "magma")

g4
```


```{r}
acro_new = acro_scores %>% 
  filter(plate_position != c("Plate7_D006", "Plate6_E009", "Plate6_G009", "Plate6_D012", "Plate7_H009", "Plate7_A010", "Plate6_B012"))


ord_acro
```


Variation partitioning based on dbRDA results. 
```{r}
varp_acro = varpart(dist_acro, ~ Depth, ~DHW6, ~GPS_S, ~recent.maxDHW,  data = adonis_meta_acro)

plot(varp_acro, digits = 2, Xnames = c("Depth", "DHW6", "GPS_S", "recent.maxDHW"), bg = c("navy", "tomato", "orange", "green"))
```
**Recent max DHW is explaining the most variability in symbiont communities on it's own (4.1% of variability). In contrast, depth seems to explain very little in terms of symbiont communities (0.7%). ** 

**The lack of influence of depth could be attributed to the nutrient poor waters of the Coral Sea and light availability compared to that of the GBR or other regions where depth maintains a greater influence ** 

Locally driven (acro) v. connectivity driven (poci)

```{r}
4.1/0.7
```


```{r}
library(ggpubr)
Fig3 = ggarrange (g1 + g2 + g3 + g4, nrow = 1) + 
    theme_classic() +
  theme(axis.line = element_blank(),
        panel.border = element_rect(size = .5, fill = "transparent"),
        legend.title = element_blank(),
        legend.background = element_rect(fill = "transparent"),
        legend.position = c(0.5, 1),
        legend.text = element_text(face = "italic", size = 7, family = "Helvetica"),
        legend.key.size = unit(2, units = "mm"),
        legend.margin = margin(0,0,0,0, unit = "mm"),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 8, family = "Helvetica"),
        axis.title = element_text(size = 9, family = "Helvetica"))  
    
Fig3


#change so it's just one row
#remove x and y axis labels
#make colour scheme stay the same for each reef across 4 host species
```


```{r}
save(adonis_meta_acro, file = "Data/adonis_meta_acro.Rdata")
save(adonis_meta_pver, file = "Data/adonis_meta_pver.Rdata")
save(adonis_meta_pmea, file = "Data/adonis_meta_pmea.Rdata")
save(adonis_meta_punk, file = "Data/adonis_meta_punk.Rdata")

```


**Significant factors for symbionts within each host species ** 
P. verrucosa: Depth *, DHW3 *, latitude ***
P. meandrina: DHW6 *, varSST **, latitude ***
P. unknown: varSST *, latitude ***
A. humilis: Depth *, DHW3 *, DHW6 ***, recent.maxDHW ***, latitude ***

Acropora is the less heat tolerant species. Also appears that symbiont communities are more sensitive to several enviro. factors compared to Pocillopora (more heat tolerant species)

Hierarchy of symbiont structuring as follows: host species > latitude > thermal history > depth
Symbiont species in P. meandrina and P. unknown (C. Latusorum) has consistent response to enviroment irrespective of host species. Responds to latitude and variability in SST, regardless of which host species it is living in. 

**Figures for manuscript** 
Fig 1: map (half page)
Fig 2: (full page) 3 trees + distances
- remove GPS south and replace with reef
- character matrix with species, reef and clade for groupings 
Fig 3: (half) ORDINATION plots
Fig 4: (half) tanglegram 


It seems that within hosts, the symbiont communities respond to different environmental pressures.
