==> 3_NetworkAnalysis.Rmd <==
---
title: "3_NetworkAnalysis"
author: "Magena Marzonie"
date: "15/06/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


- how do people challenge networks w variables external to the network?
- remove sequences that are everywhere. 
- keep simplifying until you have threshold where structure is clear. 

g graph. port for i graph objects straight into ggplot. Here are some tutorials for layout, etc. 
https://github.com/thomasp85/ggraph
https://www.data-imaginist.com/2017/ggraph-introduction-layouts/



How do we identify which sequences are driving these difference? 
- Best way to do presence/absence scores at reefs? 



```{r}
#load libraries 
library(ggraph)
library(igraph)
library(tidyverse)
library(tidygraph)
```


```{r}
#load sequence data
load("Data/pver_seqs.Rdata")
load("Data/pmea_seqs.Rdata")
load("Data/punk_seqs.Rdata")
load("Data/acro_seqs.Rdata")

#load environmental data
load("Data/adonis_meta_pver.Rdata")
load("Data/adonis_meta_pmea.Rdata")
load("Data/adonis_meta_punk.Rdata")
load("Data/adonis_meta_acro.Rdata")
```

```{r}
#list of sequences per reef. 
pmea_seqreef <- pmea_seqs %>% 
distinct(name, Reef)

#removing sequences that are abundant across 5 or more reefs, as these sequences will not be 'diagnostic' if present across many reefs 
remove_list <- pmea_seqreef %>% dplyr::count(name) %>% filter(n >=5) %>% mutate(name = as.character(name)) %>% pull(name)

pmea_edges <- pmea_seqreef %>%
  filter(name %in% remove_list)
```
#Two dataframes
1) node list and edge list: samples / sequences 
2) node list and edge list sequences / sequences 

```{r}
#within each reef, get the sum of sequences across all samples collected. 
pmea_reefseqs = pmea_seqs %>%  
  filter(!name %in% remove_list) %>%
  select(name, Reef, sample_name, value ) %>%
  group_by(Reef, name) %>%
  summarise(value = sum(value))  %>%
select(from = name, to = Reef, value)

pmea_reefseqs

```



```{r}
# Example data

#seqs <- paste0("seq_", seq(1, 500, 1)) # 500 unique sequence ids
#reef <- paste0("reef_", letters[1:15]) # 15 distinct reefs

#seq_reef_network <- pmea_reefseqs %>%  expand_grid(from = name, to = Reef)

#seq_reef_network <- expand_grid(from = name, to = Reef) %>% # cross all seqs with all reefs
 # group_by(to) %>%
 # mutate(freq = cur_group_id()) %>%
 # sample_n(freq[1]+6) %>% # sample a subset of total cross to create some between-reef variance in composition
  #ungroup() %>%
# mutate(count = abs(round(rnorm(n(), mean = 2000, sd = 2000), digits = 0))) # add some random count data

```

```{r}
  pmea_reefseqs %>% as_tbl_graph() %>% # convert to tidygraph format
   activate(nodes) %>%
   mutate(attribute = case_when(str_detect(name, "Bougainville|Chilcott|Flinders|Frederick|Herald|Holmes|Lihou|Marion|Moore|Osprey|Saumarez|Willis|Wreck") ~ "Reef", TRUE ~ "Sequence"), # add node attribute info
          centrality = centrality_degree()) %>% # can measure which nodes/reefs are 'central'
  ggraph(layout = "fr") +
   geom_edge_link(aes(width = value, alpha = value)) +
   scale_edge_width(range = c(0.2, 2.5)) + # control size
   geom_node_point(aes(shape = attribute, fill = centrality), size = 4) +
   scale_shape_manual(values = c(21, 22)) +
  scale_fill_viridis_c(option = "magma")#+
#  geom_text(aes(label = "Reef"), size = 4)
```
